+++ extension/tests/webcat/validators.test.ts
@@ -1,14 +1,7 @@
 // validateCSP.test.ts
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-import { WebcatErrorCode } from "../../src/webcat/interfaces/errors";
-import {
-  enforceHTTPS,
-  extractAndValidateHeaders,
-  isSafeRelativeLocation,
-  validateCSP,
-  validateProtocolAndPort,
-} from "../../src/webcat/validators";
+import { validateCSP } from "../../src/webcat/validators";
 
 // Mocks (unchanged)
 vi.mock("../../src/webcat/logger", () => ({
@@ -401,308 +394,6 @@ describe("validateCSP", () => {
 
     await expect(validateCSP(csp, trustedFQDN, valid_sources)).rejects.toThrow(
       "CSP contains a comma",
-    );
-  });
-});
-
-describe("validateProtocolAndPort", () => {
-  it("allows https with default port", () => {
-    const url = new URL("https://example.com");
-    expect(validateProtocolAndPort(url)).toBe(true);
-  });
-
-  it("allows https with explicit port 443", () => {
-    const url = new URL("https://example.com:443");
-    expect(validateProtocolAndPort(url)).toBe(true);
-  });
-
-  it("allows http with default port", () => {
-    const url = new URL("http://example.com");
-    expect(validateProtocolAndPort(url)).toBe(true);
-  });
-
-  it("allows http with explicit port 80", () => {
-    const url = new URL("http://example.com:80");
-    expect(validateProtocolAndPort(url)).toBe(true);
-  });
-
-  it("rejects https with a non-standard port", () => {
-    const url = new URL("https://example.com:8443");
-    expect(validateProtocolAndPort(url)).toBe(false);
-  });
-
-  it("rejects http with a non-standard port", () => {
-    const url = new URL("http://example.com:8080");
-    expect(validateProtocolAndPort(url)).toBe(false);
-  });
-
-  it("rejects unsupported protocol (ftp)", () => {
-    const url = new URL("ftp://example.com");
-    expect(validateProtocolAndPort(url)).toBe(false);
-  });
-
-  it("rejects unsupported protocol (file)", () => {
-    const url = new URL("file:///etc/passwd");
-    expect(validateProtocolAndPort(url)).toBe(false);
-  });
-
-  it("rejects data: URLs", () => {
-    const url = new URL("data:text/plain,hello");
-    expect(validateProtocolAndPort(url)).toBe(false);
-  });
-});
-
-describe("enforceHTTPS", () => {
-  it("redirects http to https for normal domains", () => {
-    const url = new URL("http://example.com/path");
-    const redirect = enforceHTTPS(url);
-
-    expect(redirect).toBe("https://example.com/path");
-    expect(url.protocol).toBe("https:");
-  });
-
-  it("does nothing for already-https URLs", () => {
-    const url = new URL("https://example.com/path");
-    const redirect = enforceHTTPS(url);
-
-    expect(redirect).toBeUndefined();
-    expect(url.protocol).toBe("https:");
-  });
-
-  it("does not redirect .onion domains", () => {
-    const url = new URL("http://example.onion/path");
-    const redirect = enforceHTTPS(url);
-
-    expect(redirect).toBeUndefined();
-    expect(url.protocol).toBe("http:");
-  });
-
-  it("handles localhost correctly (redirects to https)", () => {
-    const url = new URL("http://localhost:8080/path");
-    const redirect = enforceHTTPS(url);
-
-    expect(redirect).toBe("https://localhost:8080/path");
-    expect(url.protocol).toBe("https:");
-  });
-
-  it("handles subdomains correctly", () => {
-    const url = new URL("http://a.b.c.example.com/");
-    const redirect = enforceHTTPS(url);
-
-    expect(redirect).toBe("https://a.b.c.example.com/");
-  });
-
-  it("is idempotent when called twice", () => {
-    const url = new URL("http://example.com/");
-    const first = enforceHTTPS(url);
-    const second = enforceHTTPS(url);
-
-    expect(first).toBe("https://example.com/");
-    expect(second).toBeUndefined();
-    expect(url.protocol).toBe("https:");
-  });
-});
-
-describe("isSafeRelativeLocation", () => {
-  it("allows absolute-path relative locations", () => {
-    expect(isSafeRelativeLocation("/")).toBe(true);
-    expect(isSafeRelativeLocation("/login")).toBe(true);
-    expect(isSafeRelativeLocation("/a/b/c")).toBe(true);
-  });
-
-  it("allows parent-relative paths", () => {
-    expect(isSafeRelativeLocation("../login")).toBe(true);
-    expect(isSafeRelativeLocation("../a/b")).toBe(true);
-  });
-
-  it("allows same-relative paths", () => {
-    expect(isSafeRelativeLocation("./login")).toBe(true);
-  });
-
-  it("trims whitespace before validation", () => {
-    expect(isSafeRelativeLocation(" /login ")).toBe(true);
-    expect(isSafeRelativeLocation("  ../login")).toBe(true);
-  });
-
-  it("rejects protocol-relative URLs", () => {
-    expect(isSafeRelativeLocation("//evil.com")).toBe(false);
-    expect(isSafeRelativeLocation("///evil.com")).toBe(false);
-  });
-
-  it("rejects absolute URLs with schemes", () => {
-    expect(isSafeRelativeLocation("https://evil.com")).toBe(false);
-    expect(isSafeRelativeLocation("http://evil.com")).toBe(false);
-    expect(isSafeRelativeLocation("ftp://evil.com")).toBe(false);
-    expect(isSafeRelativeLocation("javascript:alert(1)")).toBe(false);
-    expect(isSafeRelativeLocation("blob:abcd")).toBe(false);
-    expect(isSafeRelativeLocation("data:text/plain,hi")).toBe(false);
-  });
-
-  it("rejects backslash-based paths", () => {
-    expect(isSafeRelativeLocation("\\evil.com")).toBe(false);
-    expect(isSafeRelativeLocation("/\\evil.com")).toBe(false);
-    expect(isSafeRelativeLocation("\\\\evil.com")).toBe(false);
-  });
-
-  it("rejects bare relative paths", () => {
-    expect(isSafeRelativeLocation("login")).toBe(false);
-  });
-
-  it("allows encoded slashes (no decoding is performed)", () => {
-    expect(isSafeRelativeLocation("/%2f%2fevil.com")).toBe(true);
-    expect(isSafeRelativeLocation("%2f%2fevil.com")).toBe(false);
-  });
-
-  it("allows control characters (current behavior)", () => {
-    expect(isSafeRelativeLocation("/foo\nbar")).toBe(true);
-    expect(isSafeRelativeLocation("/foo\rbar")).toBe(true);
-    expect(isSafeRelativeLocation("/foo\tbar")).toBe(true);
-  });
-});
-
-describe("extractAndValidateHeaders", () => {
-  it("requires CSP for non-cached responses", () => {
-    const details = {
-      responseHeaders: [{ name: "x-webcat-version", value: "1.2.3" }],
-      fromCache: false,
-    } as browser.webRequest._OnHeadersReceivedDetails;
-
-    const result = extractAndValidateHeaders(details);
-
-    expect(result).toBeInstanceOf(Error);
-    expect((result as { code: string }).code).toBe(
-      WebcatErrorCode.Headers.MISSING_CRITICAL,
-    );
-  });
-
-  it("allows missing CSP for fully cached responses", () => {
-    const details = {
-      responseHeaders: [{ name: "x-webcat-version", value: "1.2.3" }],
-      fromCache: true,
-    } as browser.webRequest._OnHeadersReceivedDetails;
-
-    const result = extractAndValidateHeaders(details);
-
-    expect(result).toBeInstanceOf(Map);
-    expect((result as Map<string, string>).has("content-security-policy")).toBe(
-      false,
-    );
-  });
-
-  it("allows missing CSP for 304 responses", () => {
-    const details = {
-      responseHeaders: [{ name: "x-webcat-version", value: "1.2.3" }],
-      fromCache: false,
-      statusCode: 304,
-    } as browser.webRequest._OnHeadersReceivedDetails;
-
-    const result = extractAndValidateHeaders(details);
-
-    expect(result).toBeInstanceOf(Map);
-    expect((result as Map<string, string>).has("content-security-policy")).toBe(
-      false,
-    );
-  });
-
-  it("blocks Location header on sub-resource requests (script)", () => {
-    const details = {
-      responseHeaders: [
-        { name: "Location", value: "/other.js" },
-        { name: "Content-Security-Policy", value: "default-src 'self'" },
-      ],
-      fromCache: false,
-      type: "script",
-    } as browser.webRequest._OnHeadersReceivedDetails;
-
-    const result = extractAndValidateHeaders(details);
-
-    expect(result).toBeInstanceOf(Error);
-    expect((result as { code: string }).code).toBe(
-      WebcatErrorCode.Headers.LOCATION_SUBRESOURCE,
-    );
-  });
-
-  it("blocks Location header on sub-resource requests (stylesheet)", () => {
-    const details = {
-      responseHeaders: [
-        { name: "Location", value: "/other.css" },
-        { name: "Content-Security-Policy", value: "default-src 'self'" },
-      ],
-      fromCache: false,
-      type: "stylesheet",
-    } as browser.webRequest._OnHeadersReceivedDetails;
-
-    const result = extractAndValidateHeaders(details);
-
-    expect(result).toBeInstanceOf(Error);
-    expect((result as { code: string }).code).toBe(
-      WebcatErrorCode.Headers.LOCATION_SUBRESOURCE,
-    );
-  });
-
-  it("blocks Location header on sub-resource requests (xmlhttprequest)", () => {
-    const details = {
-      responseHeaders: [
-        { name: "Location", value: "/api/other" },
-        { name: "Content-Security-Policy", value: "default-src 'self'" },
-      ],
-      fromCache: false,
-      type: "xmlhttprequest",
-    } as browser.webRequest._OnHeadersReceivedDetails;
-
-    const result = extractAndValidateHeaders(details);
-
-    expect(result).toBeInstanceOf(Error);
-    expect((result as { code: string }).code).toBe(
-      WebcatErrorCode.Headers.LOCATION_SUBRESOURCE,
-    );
-  });
-
-  it("allows safe relative Location header on main_frame", () => {
-    const details = {
-      responseHeaders: [
-        { name: "Location", value: "/login" },
-        { name: "Content-Security-Policy", value: "default-src 'self'" },
-      ],
-      fromCache: false,
-      type: "main_frame",
-    } as browser.webRequest._OnHeadersReceivedDetails;
-
-    const result = extractAndValidateHeaders(details);
-
-    expect(result).toBeInstanceOf(Map);
-  });
-
-  it("allows safe relative Location header on sub_frame", () => {
-    const details = {
-      responseHeaders: [
-        { name: "Location", value: "/embed" },
-        { name: "Content-Security-Policy", value: "default-src 'self'" },
-      ],
-      fromCache: false,
-      type: "sub_frame",
-    } as browser.webRequest._OnHeadersReceivedDetails;
-
-    const result = extractAndValidateHeaders(details);
-
-    expect(result).toBeInstanceOf(Map);
-  });
-
-  it("blocks external Location header even on main_frame", () => {
-    const details = {
-      responseHeaders: [
-        { name: "Location", value: "https://evil.com" },
-        { name: "Content-Security-Policy", value: "default-src 'self'" },
-      ],
-      fromCache: false,
-      type: "main_frame",
-    } as browser.webRequest._OnHeadersReceivedDetails;
-
-    const result = extractAndValidateHeaders(details);
-
-    expect(result).toBeInstanceOf(Error);
-    expect((result as { code: string }).code).toBe(
-      WebcatErrorCode.Headers.LOCATION_EXTERNAL,
     );
   });
 });
